% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP.R
\name{BP}
\alias{BP}
\title{Run a shiny application to fit bone section}
\usage{
BP()
}
\value{
Nothing
}
\description{
Run a shiny application to fit bone section
}
\details{
BP runs a shiny application to fit bone section
}
\examples{
\dontrun{
# Not run:
library(BoneProfileR)
BP()
}
}
\author{
Marc Girondot \email{marc.girondot@gmail.com}
}
