% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_Erinaceus_europaeus.R
\docType{data}
\name{Erinaceus_europaeus}
\alias{Erinaceus_europaeus}
\title{Example of hedgehog femur}
\format{
A png image.
}
\usage{
Erinaceus_europaeus
}
\description{
Example of hedgehog femur. A drawing produced by Michel Laurin.
}
\details{
Example of hedgehog femur
}
\examples{
bone <- Erinaceus_europaeus
plot(bone)
}
\seealso{
Other BoneProfileR: 
\code{\link{BP_AutoFit}()},
\code{\link{BP_ChooseBackground}()},
\code{\link{BP_ChooseCenter}()},
\code{\link{BP_ChooseForeground}()},
\code{\link{BP_DetectBackground}()},
\code{\link{BP_DetectCenters}()},
\code{\link{BP_DetectForeground}()},
\code{\link{BP_DuplicateAnalysis}()},
\code{\link{BP_EstimateCompactness}()},
\code{\link{BP_FitBayesianCompactness}()},
\code{\link{BP_FitMLCompactness}()},
\code{\link{BP_FitMLRadialCompactness}()},
\code{\link{BP_GetFittedParameters}()},
\code{\link{BP_ListAnalyses}()},
\code{\link{BP_LnLCompactness}()},
\code{\link{BP_OpenImage}()},
\code{\link{BP_Report}()},
\code{\link{plot.BoneProfileR}()},
\code{\link{summary.BoneProfileR}()}
}
\author{
Marc Girondot \email{marc.girondot@gmail.com}
}
\concept{BoneProfileR}
\keyword{femur}
\keyword{hedgehog}
